requirements.txt
streamlit==1.33.0
pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.4
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0